Create RemixingRenderer to allow playback of unusual formats
javax.sound is relatively restrictive about what sample rates and number of channels it can support. This varies from system to system.

Create class RemixingRenderer which checks whether the audio given has a sensible sample rate and number of channels. If not, it resamples and demultiplexes as needed before invoking javax.sound.

Migrate MediaPlayer to use the new RemixingRenderer.